scripts/pyproject.toml
[project]
name = "lex-harness-scripts"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fire>=0.7.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0",
    "pymupdf>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "shared", "greece", "tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --ignore=tests/integration"
markers = ["live: marks tests as requiring live internet access"]
